+++ src/emulate_files/DPReg.c
@@ -5,7 +5,7 @@ static void arithmeticDPReg(uint8_t opc, uint8_t opr, uint8_t rd, uint8_t rn, ui
     Pstate pstate = state->pstate;
 
     uint8_t shift = extractBits(opr, 1, 2);
-    uint8_t op2 = bitShift(shift, readRegister(rm, 1, generalRegisters), operand);
+    uint64_t op2 = bitShift(shift, readRegister(rm, 1, generalRegisters), operand);
 
     uint64_t valueToWrite;
     switch (opc) {
@@ -36,7 +36,7 @@ static void logicalDPReg(uint8_t opc, uint8_t opr, uint8_t rd, uint8_t rn, uint8
     uint64_t *generalRegisters = state->generalRegisters;
 
     uint8_t shift = extractBits(opr, 1, 2);
-    uint8_t op2 = bitShift(shift, readRegister(rm, 1, generalRegisters), operand);
+    uint64_t op2 = bitShift(shift, readRegister(rm, 1, generalRegisters), operand);
     bool n = opr % 2;
 
     uint64_t valueToWrite;
