+++ src/emulate_files/definitions.h
@@ -23,8 +23,8 @@ typedef struct {
 typedef struct {
     Pstate pstate;
     uint64_t generalRegisters[NUM_REGISTERS];
-    uint32_t currAddress; // hexadecimal address to represents PC
-    uint32_t *memory;
+    uint64_t currAddress; // hexadecimal address to represents PC
+    uint64_t *memory;
 } state;
 
 #endif
